workaround for git reset issues on windows + upgrading jgit + adding a sample to show how to use tsm with mvn exec plugin

+++ tsm/src/main/java/com/tomitribe/tsm/file/TempDir.java
@@ -25,7 +25,9 @@ public interface TempDir {
         Runtime.getRuntime().addShutdownHook(new Thread() {
             @Override
             public void run() {
-                Files.remove(dir);
+                if (dir.isDirectory()) {
+                    Files.remove(dir);
+                }
             }
         });
         return workDir;
